+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/open-ghost-overlay/main.js
@@ -11,13 +11,23 @@ define(function() {
 
     'use strict';
 
-    var templates = {
+    var translations = {},
+        defaultTranslations = {
+            info: 'content.contents.settings.copy-locale.info',
+            title: 'content.contents.settings.copy-locales.title',
+            selectDefault: 'content.contents.settings.copy-locale.select-default',
+            new: 'content.contents.settings.copy-locale.new',
+            copy: 'content.contents.settings.copy-locale.copy',
+            ok: 'content.contents.settings.copy-locale-overlay.ok'
+        },
+
+        templates = {
             openGhost: function() {
                 return [
                     '<div class="copy-locale-overlay-content grid">',
                     '   <div class="grid-row">',
                     '       <p class="info">',
-                    this.sandbox.translate('content.contents.settings.copy-locale.info'),
+                    getTranslation.call(this, 'info'),
                     '       </p>',
                     '   </div>',
                     '   <div class="grid-row">',
@@ -26,7 +36,7 @@ define(function() {
                     '           <span class="icon"></span>',
                     '       </div>',
                     '       <label for="copy-locale-new">',
-                    this.sandbox.translate('content.contents.settings.copy-locale.new'),
+                    getTranslation.call(this, 'new'),
                     '       </label>',
                     '   </div>',
                     '   <div class="grid-row">',
@@ -35,7 +45,7 @@ define(function() {
                     '           <span class="icon"></span>',
                     '       </div>',
                     '       <label for="copy-locale-copy">',
-                    this.sandbox.translate('content.contents.settings.copy-locale.copy'),
+                    getTranslation.call(this, 'copy'),
                     '       </label>',
                     '       <div id="copy-locale-overlay-select" />',
                     '   </div>',
@@ -45,15 +55,26 @@ define(function() {
         },
 
         /**
+         * Retrieves translation for given identifier.
+         *
+         * @param {String} id
+         *
+         * @return {String}
+         */
+        getTranslation = function(id) {
+            return this.sandbox.translate(translations[id]);
+        },
+
+        /**
          * start a new overlay
-         * @param {String} titleKey translation key
+         * @param {String} title the title of the overlay
          * @param {String} template template for the content
          * @param {Boolean} okButton
          * @param {undefined|String} instanceName
          * @param {undefined|function} okCallback
          * @param {undefined|function} cancelCallback
          */
-        startOverlay = function(titleKey, template, okButton, instanceName, okCallback, cancelCallback) {
+        startOverlay = function(title, template, okButton, instanceName, okCallback, cancelCallback) {
             if (!instanceName) {
                 instanceName = 'node';
             }
@@ -71,7 +92,7 @@ define(function() {
                 buttons.push({
                     type: 'ok',
                     align: 'right',
-                    text: this.sandbox.translate('content.contents.settings.' + instanceName + '.ok')
+                    text: getTranslation.call(this, 'ok')
                 });
             }
 
@@ -88,7 +109,7 @@ define(function() {
                         skin: 'medium',
                         slides: [
                             {
-                                title: this.sandbox.translate(titleKey),
+                                title: title,
                                 data: template,
                                 buttons: buttons,
                                 okCallback: okCallback,
@@ -102,17 +123,25 @@ define(function() {
 
     return {
         /**
-         * starts a copy ghost overlay
+         * Starts a copy ghost overlay.
+         *
          * @param {Object} item
+         * @param {Object} newTranslations
+         *
          * @returns {Object}
          */
-        openGhost: function(item) {
+        openGhost: function(item, newTranslations) {
             var def = this.sandbox.data.deferred();
+
+            // overwrite translations
+            translations = this.sandbox.util.extend(true, {}, defaultTranslations, newTranslations);
 
             startOverlay.call(
                 this,
-                'content.contents.settings.copy-locale.title',
-                templates.openGhost.call(this), true, 'copy-locale-overlay',
+                getTranslation.call(this, 'title'),
+                templates.openGhost.call(this),
+                true,
+                'copy-locale-overlay',
                 function() {
                     var copy = this.sandbox.dom.prop('#copy-locale-copy', 'checked'),
                         src = this.sandbox.dom.data('#copy-locale-overlay-select', 'selectionValues');
@@ -138,7 +167,7 @@ define(function() {
                     options: {
                         el: '#copy-locale-overlay-select',
                         instanceName: 'copy-locale-to',
-                        defaultLabel: this.sandbox.translate('content.contents.settings.copy-locale.select-default'),
+                        defaultLabel: getTranslation.call(this, 'selectDefault'),
                         data: item.concreteLanguages
                     }
                 }